reorganized test resources, made ParsingTester more generic
