+++ src/test/java/com/digitalpebble/storm/crawler/TestOutputCollector.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.digitalpebble.storm.crawler.parse.filter;
+package com.digitalpebble.storm.crawler;
 
 import java.util.ArrayList;
 import java.util.Collection;
